# perf: speed up stringify_string