--- a/package.json
+++ b/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "devalue",
+	"name": "devalue2",
 	"description": "Gets the job done when JSON.stringify can't",
 	"version": "4.3.0",
 	"repository": "Rich-Harris/devalue",