--- a/package.json
+++ b/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "devalue2",
+	"name": "devalue",
 	"description": "Gets the job done when JSON.stringify can't",
 	"version": "4.3.0",
 	"repository": "Rich-Harris/devalue",